save working state
