+++ dashboard.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"encoding/json"
+	"fmt"
 	"net/http"
 )
 
@@ -9,25 +10,29 @@ type DashboardInfo struct {
 	UserCount int `json:"userCount"`
 }
 
-func countUsers(accId string) int {
-	metricDB := GetDB()
-	c := metricDB.countAllUniqueUsersInAccount(accId)
-	return c
-}
-
 func DashboardHandler(w http.ResponseWriter, r *http.Request) {
-	acc := findUserAccountFromActiveToken(r)
-	acc.Users = countUsers(acc.AccountId)
-	setAccountInfo(acc)
+	acc, found := findUserAccountFromActiveToken(r)
+	if !found {
+		w.WriteHeader(http.StatusNotFound)
+		return
+	}
+	if acc.Users != acc.CountAssociatedUsers() {
+		fmt.Println(acc.Users)
+		acc.Users = acc.CountAssociatedUsers()
+		acc.UpdateSelf()
+
+		fmt.Println(acc.Users)
+	}
 
 	dasboardInfo := DashboardInfo{
 		UserCount: acc.Users,
 	}
 
 	w.Header().Set(contentTypeHeader, contentTypeJSON)
-	w.WriteHeader(http.StatusOK)
 	resJSON, err := json.Marshal(dasboardInfo)
 	if err == nil {
 		w.Write(resJSON)
+		return
 	}
+	w.WriteHeader(http.StatusInternalServerError)
 }
